Fix MSVC build errors and save dependency cache before build
- Replace POSIX-only gmtime/timegm/strptime with portable variants
  (gmtime_s/_mkgmtime and a portable strptime shim) for MSVC.
- Use uniform_int_distribution<int> instead of uint8_t, which MSVC's
  <random> rejects.
- Add explicit casts for size_t->int conversions flagged by MSVC C4267,
  including a portable HMAC data-length macro for the differing
  size_t/int OpenSSL signatures.
- Split the Conan CI cache into restore + explicit save right after
  dependency provisioning so a later build failure does not lose
  dependencies.

+++ src/cipher.cc
@@ -166,7 +166,7 @@ void encrypt_cbc(std::istream &src, std::ostream &dst,
     std::array<uint8_t, 16> src_block{}, dst_block{};
     std::array<uint8_t, 16> src_block_xor_iv{};
 
-    std::fill(src_block.begin(), src_block.end(), 16);
+    std::fill(src_block.begin(), src_block.end(), static_cast<uint8_t>(16));
     std::transform(src_block.begin(), src_block.end(), prv.begin(),
                    src_block_xor_iv.begin(), std::bit_xor<uint8_t>());
